% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.specific.R
\docType{methods}
\name{deg.specific}
\alias{deg.specific}
\title{Predict the patient-specific DEGs using bi-clustering analysis}
\usage{
deg.specific(deg, test.patients = NULL, min.genes = 50, min.patients = 5,
  overlap = 0.85, cores = 1)
}
\arguments{
\item{deg}{a 'deg' object. This should be the output of \code{\link{bi.deg}}}

\item{test.patients}{the patients to test. Only the patients in 'test.patients' are used as seed for bi-clustering analysis}

\item{min.genes}{the minimum number of genes}

\item{min.patients}{the minimum number of patients. It includes the patients as seeds  (see details).}

\item{overlap}{the minimum similarity for selected DEGs from two or more patients}

\item{cores}{the thread number}
}
\value{
A 'deg.specific' or 'deg.specific.test' object. It has a key of "decd.input", which stores the binary DEG matrix, genes, patients and used parameter setting. Other keys are the patients IDs and they store the cross-validated DEGs.
}
\description{
This function is use the output of \code{\link{bi.deg}} as input to predict the differentially expressed genes (DEGs) for each patient by cross-validation of multiple patients.
}
\details{
The DEGs from \code{\link{bi.deg}} are mixed with noises, e.g. the DEGs not associated with disease. This is especially true when the differential expression analysis tests are done using one variable again references. The assumption behind his analysis is that the disease associated DEGs will be observed in other patients. This function implements a bi-clustering algorithm to find the DEGs shared by 'min.patients' in the binary DEG matrix. In this process, each patient is used as seed and its DEGs are gradually excluded to find if there is a DEG list which is observed in 'min.patients' when the similarity is greater is 'overlap'.

'test.patients' option is used to find the cross-validated DEGs for some interest patients. Otherwise, all the patients will be used as seeds. 'overlap' is the threshold to determine the minimum similarity between neighbor and seed patients.
}
\examples{
\dontrun{
# the DEGs has at least 100 genes and validated by 5 other patients
res.deg <- deg.specific(deg, min.genes=100, min.patients=5, overlap=0.85)
}
}
\author{
Guofeng Meng
}
\references{

}

