% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCurve.R
\docType{methods}
\name{module.curve}
\alias{module.curve}
\title{Display the patient-gene numbers during bi-clustering analysis}
\usage{
module.curve(res.module, mod = names(res.module)[1])
}
\arguments{
\item{res.module}{a 'seed.module' or 'cluster.module' object returned by \code{\link{seed.module}} or \code{\link{cluster.module}}}

\item{mod}{the module to plot}
}
\value{
The plot for gene and patient number.
}
\description{
Plot the curve of patient-gene number during bi-clustering analysis
}
\details{
This function is used to display the patient and gene number during the bi-clustering analysis. It can be used for users to select the better gene or patient number for breakpoints.
}
\examples{
\dontrun{
module.curve(res.module, "M1")
}
}
\author{
Guofeng Meng
}
\references{

}

