% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.extract.R
\docType{methods}
\name{module.exact}
\alias{module.exact}
\alias{module.extract}
\title{Extract the patients and genes from module}
\usage{
module.extract(res.module, mod, n.patients = NULL, n.genes = NULL)
}
\arguments{
\item{res.module}{a 'seed.module' or 'cluster.module' object returned by \code{\link{seed.module}} or \code{\link{cluster.module}}}

\item{mod}{a module name to extract genes and patients}

\item{n.patients}{the patient number to return}

\item{n.genes}{the gene number to return}
}
\value{
A list for genes or patients.
}
\description{
Return the patients and genes from module under user defined setting
}
\details{
This function is used to return the patients and genes at user defined breakpoints.

Users can set `n.patients` or `n.genes` to define the break points of bi-clustering. But it is not allowed to set both value.
}
\examples{
\dontrun{
# extract the genes and patients when 200 patients are observed in M1.
page=module.extract(res.module, "M1", n.patients=200)
head(page$patients)
head(page$geness)
# find genes and patients in patient-seeded module
module.extract(res.module$decd.specific, names(res.module$decd.specific)[1], n.patients=200)
}
}
\author{
Guofeng Meng
}
\references{

}

