% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.module.R
\docType{methods}
\name{seed.module}
\alias{seed.module}
\title{Predict the DEGs modules shared by patients}
\usage{
seed.module(deg, res.deg = NULL, test.patients = NULL, min.genes = 100,
  min.patients = 25, overlap = 0.85, model.method = c("slope.clustering",
  "max.square", "min.slope", "min.similarity")[1], cores = 1)
}
\arguments{
\item{deg}{a binary matrix. This should be the output of \code{\link{bi.deg}} or a binary matrix}

\item{res.deg}{a 'deg.specific' object. It should be the output of \code{\link{deg.specific}}. It is optional.}

\item{test.patients}{the patient IDs used as test seed to find the modules}

\item{min.genes}{the minimum number of genes in the modules}

\item{min.patients}{the minimum number of patients in the modules}

\item{overlap}{the minimum similarity for selected DEGs from two or more patients}

\item{model.method}{the method to find the breakpoint of bi-clustering. It is accepted value including 'slope.clustering', 'max.square', 'min.slope', 'min.similarity'}

\item{cores}{the thread number}
}
\value{
A seed.module object.
It has one key with prefix of "decd":

"decd.input", the input information, including binary DEG matrix,  test.patients and other parameter setting.

It may have one key of "M0":

"M0", a modules shared by all the patients. In many cases, M0 is NULL when M0 is not predicted.

Other keys are patient IDs, which are the modules predicted with DEG seed of the patient. Each one have several keys:

"curve", the patient-gene number during bi-clustering analysis;

"max.genes", the patient and genes when 'min.patients' is observed in bi-clustering analysis;

"max.patients", the patient and genes when 'min.genes' is reached in bi-clustering analysis";

"model", the patient and genes at the breakpoint of the `curve`;

"genes.removed", the ordered genes that are removed from module during bi-clustering analysis;

"patients.added", the ordered patients that are added to module during bi-clustering analysis
}
\description{
This function uses the output of \code{\link{bi.deg}} as input to predict the patient-DEG lists (or modules) shared by patients.
}
\details{
The function is to find the DEGs lists shared by patients. Like \code{\link{deg.specific}}, it carries out the bi-clustering analysis to the output of \code{\link{bi.deg}}. The difference is that this function has more complex setting and steps to predict DEGs modules shared by patients.

No matter whatever is the parameter setting, 'deg.module' will firstly try to find a modules shared by all the patients, where the finally patient number may less than the 'min.patients' and gene number may be less than 'min.genes'. If such module exists, it will named as 'M0' and the module genes of 'M0' will be filtered and they will not be included in other modules. Then, the patients of 'deg' will used as seed to do bi-clustering analysis.

The bi-clustering analysis is started by a DEG seed, composing of the DEGs of a patient. If 'res.deg' is set, only the patients with cross-validated DEGs will be used as seeds and the seed will be initilized with the cross-validated DEGs. Otherwise, all the patients will be used and all the DEGs are used as seed. The DEGs of patients will be gradually removed to ckeck if the left seed are observed in 'min.patients' when keeping the similarity is not less than 'overlap'. 'deg.module' will record the track of gene-patient number in the bi-clustering analysis, which is stored in 'curve' for each patient.

During the bi-clustering analysis, 'deg.module' will record the bi-clustering results at three scenarios:

  'max.genes' records the patient and genes information when the seed is observed in 'min.patient'.

  'max.patients' stores the patient and gene information when 'min.genes' are observed, which is also the terminated point of bi-clustering analysis.

  'model' stores the gene/patient information when the gene-patients number 'curve' fits the criteria of 'model.method'.

The detailed information of the bi-clustering analysis results for all used patients is stored in "decd.specific" of output list.

In this version, 'model.method' has four possible values: "slope.clustering", "max.square", "min.slope" and "min.similarity", which indicate the different four different modelling methods:

 'slope.clustering' has maximum slope changes, which may indicate the inclusion/exclusion of molecular mechanism;

 'max.square' is the gene-patients number that has the maximum product;

 'min.slope' has the minimum slope in gene-patient number curve;

 'min.similarity' is based on the similarity scores and the point with minimum similarity scores is choosed.
}
\examples{
\dontrun{
seed.mods1 <- seed.module(deg, res.deg, min.genes=100, min.patient=50, overlap=0.85,
                         model.method='slope.clustering')
seed.mods2 <- seed.module(deg, model.method='min.similarity')
}
}
\author{
Guofeng Meng
}
\references{

}

