% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.module.R
\name{summarize.cluster.module.test}
\alias{summarize.cluster.module.test}
\title{Summarize the DEG modules}
\usage{
\method{summarize}{cluster.module.test}(res.module, max.n = 10)
}
\arguments{
\item{res.module}{a 'cluster.module' object returned by \code{\link{cluster.module}}}

\item{max.n}{the number of modules to display}
}
\value{
A data.frame.
}
\description{
Summarize the DEG modules
}
\details{
This function summarize the DEG modules
}
\examples{
\dontrun{
report.module <- summarize(res.module)
}
}
\author{
Guofeng Meng
}
\references{

}

