% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.deg.R
\name{summarize.deg.specific}
\alias{summarize.deg.specific}
\title{Summarize the patient-specific DEGs}
\usage{
\method{summarize}{deg.specific}(res.deg, max.n = 10)
}
\arguments{
\item{res.deg}{a list returned by \code{\link{deg.specific}}}

\item{max.n}{the number of DEGs to display}
}
\value{
A data.frame.
}
\description{
Summarize the binary differential expression matrix transformed by \code{\link{bi.deg}} or the cross-validated DEGs predicted by \code{\link{deg.specific}}.
}
\details{
This function summarizes the gene number before and after cross-validation.
}
\examples{
\dontrun{
report.deg <- summarize.deg(res.deg)
}
}
\author{
Guofeng Meng
}
\references{

}

